@@ -5,7 +5,7 @@ author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.17.0 <2.0.0"
 
 dependencies:
@@ -24,6 +24,13 @@ dependencies:
 
   flutter_effect_kit:
     path: ../flutter_effect_kit/
+
+  app_buried_point:
+    path: ../app_buried_point/
+
+  flutter_network:  # 网络 HttpStatusCode
+    path: ../flutter_network/
+
   
 
 dev_dependencies:
