@@ -1,19 +1,26 @@
+/*
+ * @Author: dvlproad
+ * @Date: 2023-02-08 13:12:39
+ * @LastEditors: dvlproad
+ * @LastEditTime: 2023-03-19 13:12:27
+ * @Description: 
+ */
 // 图片+文字:图片在上，文字显示在图片下边
 import 'package:flutter/material.dart';
 
 class EmptyWithImageAboveTextWidget extends StatefulWidget {
-  final Color color; // 背景颜色，用于有些页面有设置背景图片，需要对此视图设成透明色
+  final Color? color; // 背景颜色，用于有些页面有设置背景图片，需要对此视图设成透明色
   final ImageProvider image;
   // image: AssetImage('assets/images/emptyview/pic_搜索为空页面.png'),
   // image: NetworkImage('https://dss1.bdstatic.com/70cFuXSh_Q1YnxGkpoWK1HF6hhy/it/u=3238317745,514710292&fm=26&gp=0.jpg'),
   final String mainTitle;
   final String subTitle;
-  final Widget buttonWidget;
+  final Widget? buttonWidget;
 
   const EmptyWithImageAboveTextWidget({
-    Key key,
+    Key? key,
     this.color,
-    this.image,
+    required this.image,
     this.mainTitle = '很抱歉，您暂无相关数据',
     this.subTitle = '去其他逛逛吧',
     this.buttonWidget,
@@ -41,34 +48,32 @@ class _EmptyWithImageAboveTextWidgetState
               'assets/empty/empty_aboveText_default.png',
               package: 'flutter_effect',
             ),
-        width: 170,
-        height: 133,
+        width: 130,
+        height: 102,
         fit: BoxFit.cover,
       ),
-      Container(height: 7),
+      Container(height: 24),
       Text(
         widget.mainTitle,
         style: TextStyle(
-          color: Color(0xff9a9a9a),
-          fontFamily: 'PingFang SC',
-          fontWeight: FontWeight.bold,
-          fontSize: 15,
-        ),
-      ),
-      Text(
-        widget.subTitle,
-        style: TextStyle(
-          color: Color(0xff9a9a9a),
-          fontFamily: 'PingFang SC',
-          fontWeight: FontWeight.w400,
+          color: Color(0xffafafaf),
           fontSize: 13,
         ),
       ),
-      Container(height: 25),
+      // Text(
+      //   widget.subTitle,
+      //   style: TextStyle(
+      //     color: Color(0xff9a9a9a),
+      //     fontFamily: 'PingFang SC',
+      //     fontWeight: FontWeight.w400,
+      //     fontSize: 13,
+      //   ),
+      // ),
+      Container(height: 28),
     ];
 
     if (widget.buttonWidget != null) {
-      columnWidgets.add(widget.buttonWidget);
+      columnWidgets.add(widget.buttonWidget!);
     } else {
       columnWidgets.add(Container(height: 35));
     }
