@@ -2,7 +2,7 @@
  * @Author: dvlproad
  * @Date: 2021-04-12 23:04:04
  * @LastEditors: dvlproad
- * @LastEditTime: 2022-05-13 16:03:56
+ * @LastEditTime: 2023-03-17 17:15:57
  * @Description: 页面框架
  */
 library flutter_effect;
@@ -34,3 +34,5 @@ export './src/basepage_extends/default_page.dart';
 
 export 'package:flutter_overlay_kit/flutter_overlay_kit.dart';
 export 'package:flutter_effect_kit/flutter_effect_kit.dart';
+
+// export 'package:flutter_theme_helper/flutter_theme_helper.dart';
