@@ -2,7 +2,7 @@
  * @Author: dvlproad
  * @Date: 2022-05-12 19:40:05
  * @LastEditors: dvlproad
- * @LastEditTime: 2022-05-13 16:22:15
+ * @LastEditTime: 2023-03-19 13:07:59
  * @Description: 堆栈信息获取工具
  */
 import 'package:stack_trace/stack_trace.dart';
@@ -11,11 +11,14 @@ printStack() {
   print('$__CLASS__ $__METHOD__($__FILE__:$__LINE__)');
 }
 
-String get __CLASS__ {
+String? get __CLASS__ {
   var frames = Trace.current().frames; //调用栈
   if (frames.length > 1) {
     var member = frames[1].member;
-    var parts = member.split(".");
+    if (member == null) {
+      return null;
+    }
+    var parts = member!.split(".");
     if (parts.length > 1) {
       return parts[0];
     }
@@ -24,11 +27,14 @@ String get __CLASS__ {
   return null;
 }
 
-String get __METHOD__ {
+String? get __METHOD__ {
   var frames = Trace.current().frames;
   if (frames.length > 1) {
     var member = frames[1].member;
-    var parts = member.split(".");
+    if (member == null) {
+      return null;
+    }
+    var parts = member!.split(".");
     if (parts.length > 1) {
       return parts[1];
     }
@@ -37,7 +43,7 @@ String get __METHOD__ {
   return null;
 }
 
-String get __FILE__ {
+String? get __FILE__ {
   var frames = Trace.current().frames;
   if (frames.length > 1) {
     return frames[1].uri.path;
@@ -46,7 +52,7 @@ String get __FILE__ {
   return null;
 }
 
-int get __LINE__ {
+int? get __LINE__ {
   var frames = Trace.current().frames;
   if (frames.length > 1) {
     return frames[1].line;
